Destroy dropdown on close
